Fixing handling of grant/revoke for materialized views and external tables
