+++ snowflake/repository_test.go
@@ -14,13 +14,7 @@ func TestSchemaQuery(t *testing.T) {
 func TestTableQuery(t *testing.T) {
 	databaseName := "db🫘"
 	schemaName := "🛟schema"
-	assert.Equal(t, `SELECT * FROM "db🫘".INFORMATION_SCHEMA.TABLES WHERE TABLE_TYPE != 'VIEW' AND TABLE_SCHEMA = '🛟schema'`, getTablesInDatabaseQuery(databaseName, schemaName))
-}
-
-func TestViewsQuery(t *testing.T) {
-	databaseName := "db🫘"
-	schemaName := "🛟schema"
-	assert.Equal(t, `SELECT * FROM "db🫘".INFORMATION_SCHEMA.VIEWS  WHERE TABLE_SCHEMA = '🛟schema'`, getViewsInDatabaseQuery(databaseName, schemaName))
+	assert.Equal(t, `SELECT * FROM "db🫘".INFORMATION_SCHEMA.TABLES WHERE TABLE_SCHEMA = '🛟schema'`, getTablesInDatabaseQuery(databaseName, schemaName))
 }
 
 func TestColumnsQuery(t *testing.T) {
